Sweeps is the judge of how many iterations
Previously I would have just done a stupid while several < enough, and
this would rarely conduct a whole number of sweeps which is silly. So
instead i’m made it to be, while sweeps < 10.
I’m yet to find out the relation for how temperature and size affect
how many sweeps need to be done.